+++ app/src/main/java/com/terra/planet/MainActivity.kt
@@ -750,6 +750,6 @@ class MainActivity : Activity() {
     }
 
     companion object {
-        const val VERSION = "0.12.3"
+        const val VERSION = "0.13.0"
     }
 }